Moving plugins, adding Keyboard integration
