+++ src/plugins/hash.js
