+++ src/plugins/navigation.js
