+++ src/plugins/plugins.js
@@ -1,7 +1,9 @@
 import Navigation from './navigation';
 import Hash from './hash';
+import Keyboard from './keyboard';
 
 export default {
   Navigation,
-  Hash
+  Hash,
+  Keyboard
 };
