+++ src/utils/keys.js
@@ -1,0 +1,10 @@
+const Keys = {
+  ENTER: 13,
+  SPACE: 32,
+  LEFT: 37,
+  UP: 38,
+  RIGHT: 39,
+  DOWN: 40
+};
+
+export default Keys;
